+++ CHANGELOG.md
@@ -18,6 +18,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Don't include invoice and subscription counts in `instance-report` output if Monero intergration is not enabled.
 - Renamed `instance_uri` configuration parameter to `instance_url` (`instance_uri` is treated as alias).
 - Log CORS errors.
+- Return `413 Payload Too Large` if media uploaded by portable actor is too large.
 
 ### Fixed